requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.52.0
django-countries==7.5
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-phonenumber-field==7.0.2
idna==3.4
MarkupSafe==2.1.2
oauthlib==3.2.2
phonenumbers==8.13.4
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-http-client==3.3.7
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sendgrid==6.9.7
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
